# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "StepZero API",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/profile/me": {
      "get": {
        "tags": [
          "profile"
        ],
        "summary": "내 프로필 조회",
        "description": "현재 로그인한 사용자의 프로필 정보를 조회합니다.",
        "operationId": "get_my_profile_api_v1_profile_me_get",
        "responses": {
          "200": {
            "description": "프로필 상세 정보와 완성도 정보를 반환합니다.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserProfileRead"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      },
      "put": {
        "tags": [
          "profile"
        ],
        "summary": "내 프로필 수정",
        "description": "현재 로그인한 사용자의 프로필 정보를 수정합니다.",
        "operationId": "update_my_profile_api_v1_profile_me_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserProfileUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "수정된 최신 프로필 정보를 반환합니다.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserProfileRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/v1/profile/me/image": {
      "post": {
        "tags": [
          "profile"
        ],
        "summary": "프로필 이미지 업로드",
        "description": "현재 로그인한 사용자의 프로필 이미지를 업로드/교체합니다.",
        "operationId": "upload_my_profile_image_api_v1_profile_me_image_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_my_profile_image_api_v1_profile_me_image_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "이미지 반영 후 최신 프로필 정보를 반환합니다.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserProfileRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
...
    "/api/v1/ops/announcements": {
      "get": {
        "tags": [
          "ops"
        ],
        "summary": "운영 공지 목록 조회",
        "description": "운영 공지 작성/수정/게시 관리를 위한 목록을 조회합니다.",
        "operationId": "get_announcements_api_v1_ops_announcements_get",
        "responses": {
          "200": {
            "description": "운영 공지 목록을 반환합니다.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnnouncementList"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      },
      "post": {
        "tags": [
          "ops"
        ],
        "summary": "운영 공지 생성",
        "description": "운영 공지를 draft 상태로 생성하고 감사로그를 남깁니다.",
        "operationId": "create_ops_announcement_api_v1_ops_announcements_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AnnouncementCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "생성된 운영 공지를 반환합니다.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnnouncementItem"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/v1/ops/announcements/{announcement_id}": {
      "patch": {
        "tags": [
          "ops"
        ],
        "summary": "운영 공지 수정",
        "description": "운영 공지 제목/내용을 수정하고 감사로그를 남깁니다.",
        "operationId": "update_ops_announcement_api_v1_ops_announcements__announcement_id__patch",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "announcement_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "description": "수정할 공지 ID",
              "title": "Announcement Id"
            },
            "description": "수정할 공지 ID"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AnnouncementUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "수정된 운영 공지를 반환합니다.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnnouncementItem"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/ops/announcements/{announcement_id}/status": {
      "patch": {
        "tags": [
          "ops"
        ],
        "summary": "운영 공지 상태 변경",
        "description": "운영 공지 상태(draft/published/archived)를 변경하고 감사로그를 남깁니다.",
        "operationId": "update_ops_announcement_status_api_v1_ops_announcements__announcement_id__status_patch",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "announcement_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "description": "상태를 변경할 공지 ID",
              "title": "Announcement Id"
            },
            "description": "상태를 변경할 공지 ID"
          },
          {
            "name": "status",
            "in": "query",
            "required": true,
            "schema": {
              "enum": [
                "draft",
                "published",
                "archived"
              ],
              "type": "string",
              "title": "Status"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "상태 변경된 운영 공지를 반환합니다.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnnouncementItem"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AnnouncementCreate": {
        "properties": {
          "title": {
            "type": "string",
            "title": "Title"
          },
          "content": {
            "type": "string",
            "title": "Content"
          }
        },
        "type": "object",
        "required": [
          "title",
          "content"
        ],
        "title": "AnnouncementCreate"
      },
      "AnnouncementItem": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "content": {
            "type": "string",
            "title": "Content"
          },
          "status": {
            "type": "string",
            "enum": [
              "draft",
              "published",
              "archived"
            ],
            "title": "Status"
          },
          "created_by": {
            "type": "integer",
            "title": "Created By"
          },
          "updated_by": {
            "type": "integer",
            "title": "Updated By"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "title",
          "content",
          "status",
          "created_by",
          "updated_by",
          "created_at",
          "updated_at"
        ],
        "title": "AnnouncementItem"
      },
      "AnnouncementList": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/AnnouncementItem"
            },
            "type": "array",
            "title": "Items"
          }
        },
        "type": "object",
        "required": [
          "items"
        ],
        "title": "AnnouncementList"
      },
      "AnnouncementUpdate": {
        "properties": {
          "title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          },
          "content": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Content"
          }
        },
        "type": "object",
        "title": "AnnouncementUpdate"
      },
...
      "Body_upload_my_profile_image_api_v1_profile_me_image_post": {
        "properties": {
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File",
            "description": "업로드할 프로필 이미지 파일"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_upload_my_profile_image_api_v1_profile_me_image_post"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "UserProfileRead": {
        "properties": {
          "nickname": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Nickname"
          },
          "profile_img": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Profile Img",
            "default": "default.png"
          },
          "is_public": {
            "type": "boolean",
            "title": "Is Public",
            "default": true
          },
          "category": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Category"
          },
          "region": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Region"
          },
          "philosophy": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Philosophy"
          },
          "experiences": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Experiences"
          },
          "awards": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Awards"
          },
          "certificates": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Certificates"
          },
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "user_id": {
            "type": "integer",
            "title": "User Id"
          },
          "updated_at": {
            "type": "string",
            "title": "Updated At"
          },
          "completeness_rate": {
            "type": "integer",
            "title": "Completeness Rate"
          },
          "full_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Full Name"
          },
          "email": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Email"
          }
        },
        "type": "object",
        "required": [
          "id",
          "user_id",
          "updated_at",
          "completeness_rate"
        ],
        "title": "UserProfileRead"
      },
      "UserProfileUpdate": {
        "properties": {
          "nickname": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Nickname"
          },
          "profile_img": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Profile Img",
            "default": "default.png"
          },
          "is_public": {
            "type": "boolean",
            "title": "Is Public",
            "default": true
          },
          "category": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Category"
          },
          "region": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Region"
          },
          "philosophy": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Philosophy"
          },
          "experiences": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Experiences"
          },
          "awards": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Awards"
          },
          "certificates": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Certificates"
          },
          "full_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Full Name"
          }
        },
        "type": "object",
        "title": "UserProfileUpdate"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/api/v1/auth/login"
          }
        }
      }
    }
  },
...
}